Update locking system
Rather than directly using a ReentrantLock, I introduce a Lock class
whose lock method is interruptible. I made this change because sometimes
I ran into problems where one thread would interrupt a thread that was
blocked on lock acquisition and because it was non-interruptible, the
thread would not stop and eventually a downstream error would crop up.
To avoid this, the new Lock is interruptible and returns true if it
acquires the lock, false otherwise.

Most of this commit is just white space because I had to move a lot of
code inside of if (lock.lock()) { ... }. Indeed, the commit is far more
readable with the '-w' flag.

+++ files/js/src/main/scala/com/swoval/files/Observers.scala
@@ -1,5 +1,6 @@
 package com.swoval.files
 
+import com.swoval.concurrent.Lock
 import com.swoval.files.Directory.Entry
 import com.swoval.files.Directory.Observer
 import com.swoval.files.Directory.OnChange
@@ -72,57 +73,63 @@ private[files] class Observers[T] extends Observer[T] with AutoCloseable {
 
   private val counter: AtomicInteger = new AtomicInteger(0)
 
-  private val lock: AnyRef = new AnyRef()
+  private val lock: Lock = new Lock()
 
   private val observers: Map[Integer, Observer[T]] = new HashMap()
 
   override def onCreate(newEntry: Entry[T]): Unit = {
     var cbs: Collection[Observer[T]] = null
-    lock.synchronized {
-      cbs = observers.values
+    if (lock.lock()) {
+      try cbs = observers.values
+      finally lock.unlock()
+      val it: Iterator[Observer[T]] = cbs.iterator()
+      while (it.hasNext) it.next().onCreate(newEntry)
     }
-    val it: Iterator[Observer[T]] = cbs.iterator()
-    while (it.hasNext) it.next().onCreate(newEntry)
   }
 
   override def onDelete(oldEntry: Entry[T]): Unit = {
     var cbs: Collection[Observer[T]] = null
-    lock.synchronized {
-      cbs = observers.values
+    if (lock.lock()) {
+      try cbs = observers.values
+      finally lock.unlock()
+      val it: Iterator[Observer[T]] = cbs.iterator()
+      while (it.hasNext) it.next().onDelete(oldEntry)
     }
-    val it: Iterator[Observer[T]] = cbs.iterator()
-    while (it.hasNext) it.next().onDelete(oldEntry)
   }
 
   override def onUpdate(oldEntry: Entry[T], newEntry: Entry[T]): Unit = {
     var cbs: Collection[Observer[T]] = null
-    lock.synchronized {
-      cbs = observers.values
+    if (lock.lock()) {
+      try cbs = observers.values
+      finally lock.unlock()
+      val it: Iterator[Observer[T]] = cbs.iterator()
+      while (it.hasNext) it.next().onUpdate(oldEntry, newEntry)
     }
-    val it: Iterator[Observer[T]] = cbs.iterator()
-    while (it.hasNext) it.next().onUpdate(oldEntry, newEntry)
   }
 
   override def onError(path: Path, exception: IOException): Unit = {
     var cbs: Collection[Observer[T]] = null
-    lock.synchronized {
-      cbs = observers.values
+    if (lock.lock()) {
+      try cbs = observers.values
+      finally lock.unlock()
+      val it: Iterator[Observer[T]] = cbs.iterator()
+      while (it.hasNext) it.next().onError(path, exception)
     }
-    val it: Iterator[Observer[T]] = cbs.iterator()
-    while (it.hasNext) it.next().onError(path, exception)
   }
 
   def addObserver(observer: Observer[T]): Int = {
     val key: Int = counter.getAndIncrement
-    lock.synchronized {
-      observers.put(key, observer)
+    if (lock.lock()) {
+      try observers.put(key, observer)
+      finally lock.unlock()
     }
     key
   }
 
   def removeObserver(handle: Int): Unit = {
-    lock.synchronized {
-      observers.remove(handle)
+    if (lock.lock()) {
+      try observers.remove(handle)
+      finally lock.unlock()
     }
   }
